Remove cast to int on Rotation and Elevation precision
They were float before, are int now. Thus, no need for casting it to int

+++ Boccia-Unity/Assets/Boccia/BCI/FanPresenter.cs
@@ -119,8 +119,8 @@ public class FanPresenter : MonoBehaviour
         if (_fineFan != null)
         {
             _fineFan.Theta = _model.GameOptions.RotationRange;
-            _fineFan.NColumns = (int)_model.GameOptions.RotationPrecision;
-            _fineFan.NRows = (int)_model.GameOptions.ElevationPrecision;
+            _fineFan.NColumns = _model.GameOptions.RotationPrecision;
+            _fineFan.NRows = _model.GameOptions.ElevationPrecision;
             _fineFan.ElevationRange = _model.GameOptions.ElevationRange;
         }
     }
